fix: update fingerprint baselines for 5.10 kernel
KVM passes through bit 27 of the `MSR_IA32_ARCH_CAPABILITIES` MSR
(0x10A) to the guest, to let them know whether the processor they're
running on is affected by RFDS. According to Intel, only Atom processors
are affected [[1]]. Our baselines for 6.1 kernel are already updated.
Now this change was ported back to 5.10, so we update baselines for it.

[1]: https://www.intel.com/content/www/us/en/developer/articles/technical/software-security-guidance/advisory-guidance/register-file-data-sampling.html

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>